+++ Exceptions/UserAlreadyExistsException.cs
@@ -2,6 +2,9 @@ using System.Runtime.Serialization;
 
 namespace PortunusAdiutor.Exceptions;
 
-class UserAlreadyExistsException : Exception
+/// <summary>
+/// 	Represents error that occur when an user that already exists is tried to be created.
+/// </summary>
+public class UserAlreadyExistsException : Exception
 {
 }
