+++ Exceptions/UserNotFoundException.cs
@@ -2,6 +2,9 @@ using System.Runtime.Serialization;
 
 namespace PortunusAdiutor.Exceptions;
 
-class UserNotFoundException : Exception
+/// <summary>
+/// 	Represents error that occur when an user tis not found.
+/// </summary>
+public class UserNotFoundException : Exception
 {
 }
