+++ Extensions/WebBuilderExtensionsMailPoster.cs
@@ -12,10 +12,25 @@ namespace PortunusAdiutor.Extensions;
 
 static public partial class WebBuilderExtensions
 {
+	/// <summary>
+	/// 	Adds <see cref="MailCodePoster{TContext, TUser, TRole, TKey, TUserClaim, TUserRole, TUserLogin, TRoleClaim, TUserToken}"/> to the <see cref="ServiceCollection"/>.
+	/// </summary>
+	/// <typeparam name="TContext"></typeparam>
+/// <typeparam name="TUser">Represents an user in the identity system.</typeparam>
+/// <typeparam name="TRole">Represents a role in the identity system.</typeparam>
+/// <typeparam name="TKey">Represents the key of an user in the identity system.</typeparam>
+/// <typeparam name="TUserClaim">Repesents a claim posessed by an user.</typeparam>
+/// <typeparam name="TUserRole">Represents the link between an user and a role.</typeparam>
+/// <typeparam name="TUserLogin">Represents a login and its associated provider for an user.</typeparam>
+/// <typeparam name="TRoleClaim">Represents a claim that is granted to all users within a role.</typeparam>
+/// <typeparam name="TUserToken">Represents an authentication token for an user.</typeparam>
+	/// <param name="builder"></param>
+	/// <param name="mailParams"></param>
 	public static void AddMailCodePoster<TContext, TUser, TRole, TKey, TUserClaim, TUserRole, TUserLogin, TRoleClaim, TUserToken>(
 		this WebApplicationBuilder builder,
 		MailCodePosterParams mailParams
-	)	where TContext : OtpIdentityDbContext<TUser, TRole, TKey, TUserClaim, TUserRole, TUserLogin, TRoleClaim, TUserToken>
+	)
+	where TContext : OtpIdentityDbContext<TUser, TRole, TKey, TUserClaim, TUserRole, TUserLogin, TRoleClaim, TUserToken>
 	where TUser : IdentityUser<TKey>, IManagedUser
 	where TRole : IdentityRole<TKey>
 	where TKey : IEquatable<TKey>
@@ -32,10 +47,27 @@ static public partial class WebBuilderExtensions
 				e.GetRequiredService<ITokenBuilder>()
 			)
 		);
-	}	public static void AddMailLinkPoster<TContext, TUser, TRole, TKey, TUserClaim, TUserRole, TUserLogin, TRoleClaim, TUserToken>(
+	}
+
+/// <summary>
+/// 	Adds <see cref="MailLinkPoster{TContext, TUser, TRole, TKey, TUserClaim, TUserRole, TUserLogin, TRoleClaim, TUserToken}"/> to the <see cref="ServiceCollection"/>.
+/// </summary>
+/// <typeparam name="TContext"></typeparam>
+/// <typeparam name="TUser"></typeparam>
+/// <typeparam name="TRole"></typeparam>
+/// <typeparam name="TKey"></typeparam>
+/// <typeparam name="TUserClaim"></typeparam>
+/// <typeparam name="TUserRole"></typeparam>
+/// <typeparam name="TUserLogin"></typeparam>
+/// <typeparam name="TRoleClaim"></typeparam>
+/// <typeparam name="TUserToken"></typeparam>
+/// <param name="builder"></param>
+/// <param name="mailParams"></param>
+	public static void AddMailLinkPoster<TContext, TUser, TRole, TKey, TUserClaim, TUserRole, TUserLogin, TRoleClaim, TUserToken>(
 		this WebApplicationBuilder builder,
 		MailLinkPosterParams mailParams
-	)	where TContext : OtpIdentityDbContext<TUser, TRole, TKey, TUserClaim, TUserRole, TUserLogin, TRoleClaim, TUserToken>
+	)
+	where TContext : OtpIdentityDbContext<TUser, TRole, TKey, TUserClaim, TUserRole, TUserLogin, TRoleClaim, TUserToken>
 	where TUser : IdentityUser<TKey>, IManagedUser
 	where TRole : IdentityRole<TKey>
 	where TKey : IEquatable<TKey>
