+++ Models/JwtCustomClaims.cs
@@ -1,6 +1,12 @@
 namespace PortunusAdiutor.Models;
 
+/// <summary>
+/// 	Class to define claims constants.
+/// </summary>
 static public class JwtCustomClaims
 {
+	/// <summary>
+	/// 	Claim of type "x-digits-code".
+	/// </summary>
 	public const string XDigitsCode = "x-digits-code";
 }
