+++ Models/Pbkdf2IdentityUser.cs
@@ -7,6 +7,10 @@ using Microsoft.AspNetCore.Identity;
 
 namespace PortunusAdiutor.Models;
 
+/// <summary>
+/// 	Implementation of <see cref="IManagedUser"/> using PBKDF2 as derivation algorithm.
+/// </summary>
+/// <typeparam name="TKey"></typeparam>
 public class Pbkdf2IdentityUser<TKey> : IdentityUser<TKey>, IManagedUser
 where TKey : IEquatable<TKey>
 {
@@ -14,22 +18,37 @@ where TKey : IEquatable<TKey>
 	private const int defaultIterCount = 262140;
 	private const int defaultHashedSize = 128;
 
+	/// <summary>
+	/// 	Initializes an instance of the class.
+	/// </summary>
+	/// <param name="email">Email of the user.</param>
+	/// <param name="salt">Salt of the user.</param>
+	/// <remarks>
+	/// 	This constructor should only be used by EF to build an object representing an existing <see cref="Pbkdf2IdentityUser{TKey}"/>.
+	/// 	</remarks>
 	public Pbkdf2IdentityUser(string email, byte[] salt)
 	{
 		Email = email;
 		Salt = salt;
 	}
 
+	/// <summary>
+	/// 	Initializes na instance of the class.
+	/// </summary>
+	/// <param name="email">Email of the user.</param>
+	/// <param name="password">Password of the user.</param>
 	public Pbkdf2IdentityUser(string email, string password)
 	{
 		Email = email;
 		SetPassword(password);
 	}
 
+	/// <inheritdoc/>
 	[MemberNotNull(nameof(Salt))]
 	public void SetPassword(string password)
 	{
-		using (var sha = SHA256.Create()) {
+		using (var sha = SHA256.Create())
+		{
 			Salt =
 				sha.ComputeHash(BitConverter.GetBytes(DateTime.UtcNow.ToBinary()));
 		}
@@ -37,6 +56,7 @@ where TKey : IEquatable<TKey>
 		return;
 	}
 
+	/// <inheritdoc/>
 	public bool ValidatePassword(string password)
 	{
 		return PasswordHash == DeriveKey(password);
@@ -53,9 +73,11 @@ where TKey : IEquatable<TKey>
 		);
 		return Convert.ToBase64String(hashed);
 	}
-	
+
+	/// <inheritdoc/>
 	public byte[] Salt { get; set; }
 
+	/// <inheritdoc/>
 	public Claim[] GetClaims()
 	{
 		var id = Id.ToString();
